requirements.txt
dnspython==2.8.0
kafka-python==2.3.0
numpy==2.4.1
pandas==2.3.3
py4j==0.10.9.9
pymongo==4.16.0
pyspark==4.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
